insert, update, delete images and refresh visual model.

+++ src/imagefactories/dbImageClass.cpp
@@ -1,6 +1,6 @@
 #include "dbImageClass.h"
 
-dbImages::dbImages(long id, QString name, const QImage& im)
+dbImages::dbImages(long long id, QString name, const QImage& im)
     : imageId(id),
     imageName(name),
     image(im) { }
